Revert "Update how prize/tier chances work"
This reverts commit c5f38eb56372228f85e24e1e2ba583de6f6bd825.

+++ paper/src/main/java/com/badbones69/crazycrates/api/PrizeManager.java
@@ -178,9 +178,9 @@ public class PrizeManager {
         if (crate.getTiers() != null && !crate.getTiers().isEmpty()) {
             for (int stopLoop = 0; stopLoop <= 100; stopLoop++) {
                 for (Tier tier : crate.getTiers()) {
-                    double chance = tier.getChance();
+                    int chance = tier.getChance();
 
-                    double num = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current().nextDouble(tier.getMaxRange()) : new Random().nextDouble(tier.getMaxRange());
+                    int num = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current().nextInt(tier.getMaxRange()) : new Random().nextInt(tier.getMaxRange());
 
                     if (num >= 1 && num <= chance) {
                         return tier;
